refactor: move FileLogger from reporting to utilities layer
Fix inverted dependency where the interfaces layer imported from the
reporting layer. FileLogger is a simple in-memory log collector with
no dependencies, so it belongs in utilities.
